Move getDataStore to the bot-wide helpers
